Update Tsubaki's GroundCloud.user.js
Essentially redid a large portion of the script. The userscript not interacts with Vue and gets and sets it's data based off that.

Still a lot of work to be done, but wanted to push this update over whatever mess there was before. It currently works just as it did before, but much quicker. And the I cleaned up the code a bit, lots of commented out stuff that needs done in a later update.